Made all changes to a batter handling of indirect jumps & insertions and also support DBT process in a splitted way so that benchs with more than 1024 instruction can be compiled in several iterations.
Features:
	-> DBT will now translate blocks of 1024 instructions. If it currently
translate those block in order, it could be easily changed but it would
need to know where to place it (and insert jump at the end of the
previous section and at the end of the concerned to makesure we do not
increment PC and go in an untranslated area).
	-> Insertions made during translation are now stored in groups
depending of the above mentionned block where they were found. During
unresolved jump resolution, only the insertions made in the block of the
destination address are considered.
	-> Resolution of destination address is now done using dichotomie
search inside the insertion list (which is sorted). This is true both
for the resolution of unresolved branches by the DBT framework and for
the resolution of indirect branch by the VLIW core.
	
=> Everything is working on benchmarks (even jpegenc).