Modified the generator to handle more code in input: Because additional profiling instructions are added to the block code, we can have situations where we have more than two instruction in a single cycle. To handle this correctly, we modified the code and now allow to load as much instruction as possible in the source VLIW instructions.
Note that we have the same limitation than before: if on the same clock
cycle we have a reference to same register it can insert false
dependencies between them. 
Still as the code will run on VLIW binaries generated by the
firstPassGenerator and with some additional instruction from profiler,
we can say that this would never happen.