New transformation: build trace
NOTE: this transformation is still under development
This transformation will merge the different block of a same procedure
in order to generate bigger blocks. Indeed, the current weakness of the
dbt process is the way control structure are handled: in code with many
small if then else, the VLIW will not be exploited at all.
The current implementation is able to merge blocks of simple if then
structures or to merge the resulting block with the next one...
This code is not tested yet...