Change to new RISC-V compiler
The new RISC-V compiler link differently and we no longer have the start
address at the begining of the text section (and no longer equal to
0x10000). Some of the previous code worked only for this specific case
and this has been corrected and updated.
The initialization will now look for the symbol _start in the elf file
in order to find out where to start the execution.
This apply for both the RISC-V binary simulator and the dbt framework.
README will be updated soon to explain how to build the new compiler...