New version which work for different (any?) vliw configuration.
If the vliw initial configuration/issue width are changed, the system
should still be able to work properly.
Benchmarks are working properly for the default configuration
(optimization level 1). 
Level two with automatic reconfiguration has been tested on matrix
multiplication only and is working.

Loop unrolling and trace building needs to be improved to work anywhere.